Add Place Type Option To Google Autocomplete Field
Added the placeType method which allows the user to pass a string of possible place values. Valid place types can be found here: https://developers.google.com/places/supported_types#table3

Also added a default type of empty string in the constructor. Passing in an empty string to Google returns all place options. This was done to override the underlying vue-google-autocomplete package that powers this field, which sets the type setting to 'address' if no type option is initially passed.